fix: harden repainting, scheduling, and compositor support code
- restore full initial damage repaint after map and unmap events
- repair fade execution, use monotonic timing, and avoid idle busy-spins
- cap painting at the configured or detected display refresh rate
- validate X properties and all numeric command-line option values
- invalidate shadows when opacity, type, frame extents, or geometry change
- avoid unnecessary second-pass painting for opaque, shadowless windows
- handle absent XRandR safely and improve root background pixmap cleanup
- make rectangle arithmetic overflow-safe and improve region handling
- add dependency-aware builds, unit tests, and an Xvfb smoke test
- update the README, man page, ignored artifacts, and fork links

+++ comp_rect.h
@@ -1,14 +1,15 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 
 typedef struct {
-    short x1;
-    short y1;
-    short x2;
-    short y2;
-    short w;
-    short h;
+    int32_t x1;
+    int32_t y1;
+    int32_t x2;
+    int32_t y2;
+    int32_t w;
+    int32_t h;
 } CompRect;
